refactor matrix functions to use float type for improved precision

+++ header/global.h
@@ -3,13 +3,13 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-void inputMatrix(vector<vector<int>> &matrix, int n, int m);
-void printMatrix(vector<vector<int>> &matrix);
-void findReducedRowEchelonForm(vector<vector<int>> &matrix);
-void findReducedRowEchelonFormRecursive(vector<vector<int>> &matrix, int r, int lead);
-void swapRows(vector<vector<int>> &matrix, int i, int j);
-void divideRow(vector<vector<int>> &matrix, int i, int val);
-void subtractRows(vector<vector<int>> &matrix, int i, int j, int lead);
-void printSolvedLinearSystem(vector<vector<int>> &matrix);
+void inputMatrix(vector<vector<float>> &matrix, int n, int m);
+void printMatrix(vector<vector<float>> &matrix);
+void findReducedRowEchelonForm(vector<vector<float>> &matrix);
+void findReducedRowEchelonFormRecursive(vector<vector<float>> &matrix, int r, int lead);
+void swapRows(vector<vector<float>> &matrix, int i, int j);
+void divideRow(vector<vector<float>> &matrix, int i, float val);
+void subtractRows(vector<vector<float>> &matrix, int i, int j, int lead);
+void printSolvedLinearSystem(vector<vector<float>> &matrix);
 
 #endif // GLOBAL_H
